add comment support to community

+++ src/main/java/com/apple/shen/community/controller/PublishController.java
@@ -1,7 +1,6 @@
 package com.apple.shen.community.controller;
 
 import com.apple.shen.community.dto.QuestionDTO;
-import com.apple.shen.community.mapper.QuestionMapper;
 import com.apple.shen.community.model.Question;
 import com.apple.shen.community.model.User;
 import com.apple.shen.community.service.QuestionService;
@@ -22,7 +21,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -42,7 +41,7 @@ public class PublishController {
             @RequestParam("title") String title,
             @RequestParam("description") String description,
             @RequestParam("tag") String tag,
-            @RequestParam("id") Integer id,
+            @RequestParam("id") Long id,
             HttpServletRequest request,
             Model model){
